pyproject.toml
[project]
name = "fastapi-startap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai>=1.61.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
